Move DB population fixtures to a `conftest.py`
