+++ datalad_registry/tests/test_new_organization/test_blueprints/test_api/test_dataset_urls.py
@@ -1,4 +1,3 @@
-from datetime import datetime
 from typing import Optional
 
 import pytest
@@ -13,130 +12,6 @@ from datalad_registry.blueprints.api.url_metadata.models import (
     URLMetadataModel,
     URLMetadataRef,
 )
-from datalad_registry.models import URL, URLMetadata, db
-
-
-@pytest.fixture
-def populate_with_2_dataset_urls(flask_app):
-    """
-    Populate the url table with 2 URLs, at position 1 and 3.
-    """
-
-    dataset_url1 = URL(url="https://example.com")
-    dataset_url2 = URL(url="https://docs.datalad.org")
-    dataset_url3 = URL(url="/foo/bar")
-
-    with flask_app.app_context():
-        for url in [dataset_url1, dataset_url2, dataset_url3]:
-            db.session.add(url)
-        db.session.commit()
-
-        db.session.delete(dataset_url2)
-        db.session.commit()
-
-
-@pytest.fixture
-def populate_with_dataset_urls(flask_app) -> list[str]:
-    """
-    Populate the url table with a list of URLs.
-
-    Returns: The list of URLs that were added to the database
-    """
-
-    urls = [
-        URL(
-            url="https://www.example.com",
-            ds_id="2a0b7b7b-a984-4c4a-844c-be3132291d7b",
-            head_describe="1234",
-            annex_key_count=20,
-            annexed_files_in_wt_count=200,
-            annexed_files_in_wt_size=1000,
-            git_objects_kb=110,
-            info_ts=datetime(2008, 7, 18, 18, 34, 32),
-            processed=True,
-            cache_path="8c8/fff/e01f2142d88690d92144b00af0",
-        ),
-        URL(
-            url="http://www.datalad.org",
-            ds_id="2a0b7b7b-a984-4c4a-844c-be3132291d7c",
-            head_describe="1234",
-            annex_key_count=40,
-            annexed_files_in_wt_count=100,
-            annexed_files_in_wt_size=400,
-            git_objects_kb=1100,
-            info_ts=datetime(2009, 6, 18, 18, 34, 32),
-            processed=True,
-            cache_path="72e/cd9/cc10534e2a9f551e32119e0e60",
-        ),
-        URL(
-            url="https://handbook.datalad.org",
-            ds_id="2a0b7b7b-a984-4c4a-844c-be3132291a7c",
-            head_describe="1234",
-            annex_key_count=90,
-            annexed_files_in_wt_count=490,
-            annexed_files_in_wt_size=1000_000,
-            git_objects_kb=4000,
-            info_ts=datetime(2004, 6, 18, 18, 34, 32),
-            processed=True,
-            cache_path="72e/4e5/4184da47e282c02ae7e568ba74",
-        ),
-        URL(
-            url="https://www.dandiarchive.org",
-            processed=False,
-            cache_path="a/b/c",
-        ),
-    ]
-
-    with flask_app.app_context():
-        for url in urls:
-            db.session.add(url)
-        db.session.commit()
-
-        return [url.url for url in urls]
-
-
-@pytest.fixture
-def populate_with_url_metadata(
-    populate_with_dataset_urls,  # noqa: U100 (unused argument)
-    flask_app,
-):
-    """
-    Populate the url_metadata table with a list of metadata
-    """
-    metadata_lst = [
-        URLMetadata(
-            dataset_describe="1234",
-            dataset_version="1.0.0",
-            extractor_name="metalad_core",
-            extractor_version="0.14.0",
-            extraction_parameter=dict(a=1, b=2),
-            extracted_metadata=dict(c=3, d=4),
-            url_id=1,
-        ),
-        URLMetadata(
-            dataset_describe="1234",
-            dataset_version="1.0.0",
-            extractor_name="metalad_studyminimet",
-            extractor_version="0.1.0",
-            extraction_parameter=dict(a=1, b=2),
-            extracted_metadata=dict(c=3, d=4),
-            url_id=1,
-        ),
-        URLMetadata(
-            dataset_describe="1234",
-            dataset_version="1.0.0",
-            extractor_name="metalad_core",
-            extractor_version="0.14.0",
-            extraction_parameter=dict(a=1, b=2),
-            extracted_metadata=dict(c=3, d=4),
-            url_id=3,
-        ),
-    ]
-
-    with flask_app.app_context():
-        for metadata in metadata_lst:
-            db.session.add(metadata)
-        db.session.commit()
 
 
 class TestCreateDatasetURL:
